requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase>=2.0.0
google-generativeai
httpx 